--- src/app/live/LiveDataUI.tsx
@@ -1,118 +1,0 @@
-'use client';
-
-import { LiveBuyBtn } from '@/components/features/live/LiveBuyBtn';
-import { LiveCalendarBtn } from '@/components/features/live/LiveCalendarBtn';
-import { LiveComment } from '@/components/features/live/LiveComment';
-import { LiveProgress } from '@/components/features/live/LiveProgress';
-import { LiveVideo } from '@/components/features/live/LiveVideo';
-import { GoBackButton } from '@/components/features/shop/ProductDetail/ProductDetailClient';
-import moment from 'moment';
-import { useState } from 'react';
-
-export const liveDummyData = [
-  {
-    id: '1',
-    start: moment('2025-07-24 00:00'),
-    end: moment('2025-07-24 18:00'),
-    title: '여름 신상 라이브 세일',
-  },
-  {
-    id: '2',
-    start: moment('2025-07-24 00:00'),
-    end: moment('2025-07-24 18:00'),
-    title: '여름  라이브 세일',
-  },
-  {
-    id: '3',
-    start: moment('2025-07-25 14:00'),
-    end: moment('2025-07-25 15:00'),
-    title: '여름  세일',
-  },
-];
-
-// 지금 방송 중인 라이브
-export const now = moment();
-export const currentLive = liveDummyData.filter(live =>
-  now.isBetween(live.start, live.end),
-);
-
-export default function LiveOverlay() {
-  /**
-   *예시)
-   *src="https://www.youtube.com/embed/VSH2d0qUkpM?si=XQ5FXY5fQjEJp-UD"
-   *src="https://www.youtube.com/embed/<여기에 있는 코드가 livePath>"
-   *src="https://www.youtube.com/embed/<여기에 있는 코드가 liveId>?si=XQ5FXY5fQjEJp-UD"
-   */
-  // 라이브 링크 아이디(비디오 가져올 때)
-  const livePath = 'hAX63N-mCxs?si=5ZoyVqA0d5n7j5bE';
-  // 라이브 아이디(댓글 가져올 때)
-  const liveId = 'yf5NOyy1SXU'; // YouTube Live Video ID
-
-  // 라이브 일정
-
-  // 오버레이 토글
-  const [showOverlay, setShowOverlay] = useState(true);
-  const handleClickOverlay = () => {
-    setShowOverlay(false);
-  };
-  return (
-    <>
-      {currentLive ? (
-        <div className="absolute top-[5%]">
-          <LiveProgress />
-        </div>
-      ) : (
-        showOverlay && (
-          <div
-            onClick={handleClickOverlay}
-            className="absolute top-0 left-0 z-5 h-[100vh] w-full bg-black/80"
-          >
-            <div className="absolute top-[50%] left-[50%] -translate-x-1/2 -translate-y-1/2 text-center text-white">
-              <p className="text-xl font-semibold">
-                현재 방송 중인 Live가 없습니다.
-              </p>
-              <p className="mt-1 text-xs font-light text-gray-200">
-                지난 방송이 궁금하다면 화면을 클릭해 보세요!
-              </p>
-            </div>
-          </div>
-        )
-      )}
-
-      {/* 라이브 캘린더 버튼 */}
-      <header className="bg-amber-200select-none absolute z-10 w-full">
-        <ul>
-          <li className="absolute top-3.5 left-0 ml-3.5 bg-amber-50">
-            <GoBackButton />
-          </li>
-          <li className="absolute top-3.5 left-[50%] translate-x-[-50%] text-xl font-bold text-white">
-            라이브
-          </li>
-          <li>
-            <LiveCalendarBtn />
-          </li>
-        </ul>
-      </header>
-
-      {/* 상품 설명, 제품 보기 링크 */}
-      <div className="absolute top-[calc(43%)] z-1 flex w-full justify-between px-3.5 text-white">
-        <span className="flex flex-col justify-end text-xs leading-loose text-[#6B7280]">
-          <p>상품 제목 타이틀</p>
-          <p>상품 설명 텍스트</p>
-        </span>
-        <span>
-          <p className="text-[40px] font-semibold text-[#FE508B]">20%</p>
-          <LiveBuyBtn />
-        </span>
-      </div>
-      {/* 라이브 비디오 */}
-      <div className="h-[60%]">
-        <LiveVideo livePath={livePath} />
-      </div>
-      {/* 라이브 댓글 */}
-      <div className="h-[48%]">
-        <LiveComment liveId={liveId} />
-      </div>
-    </>
-  );
-}
